git pull origin main

+++ PRM392_FinalProject_ProductSaleApp_Group2/app/src/main/java/com/example/prm392_finalproject_productsaleapp_group2/product/ProductDetailActivity.java
@@ -439,7 +439,7 @@ public class ProductDetailActivity extends AppCompatActivity {
                                     ? addedItem.getProduct().getProductName()
                                     : "Sản phẩm";
                             int qty = addedItem.getQuantity();
-                            int total = addedItem.getPrice();
+                            double total = addedItem.getPrice();
 
                             showToast("🛒 Đã thêm " + qty + " x " + productName +
                                     " (Tổng: " + NumberFormat.getInstance(new Locale("vi", "VN"))
